fix file submission

+++ project/backend/server/src/main/java/org/frosty/server/services/course/FileSubmissionService.java
@@ -1,9 +1,13 @@
 package org.frosty.server.services.course;
 
 import lombok.RequiredArgsConstructor;
+import org.frosty.common.exception.ExternalException;
+import org.frosty.common.response.Response;
 import org.frosty.common_service.storage.api.ObjectStorageService;
 import org.frosty.server.controller.course.FileSubmissionController;
+import org.frosty.server.entity.bo.Assignment;
 import org.frosty.server.entity.bo.FileSubmission;
+import org.frosty.server.mapper.course.AssignmentMapper;
 import org.frosty.server.mapper.course.FileSubmissionMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,18 +20,40 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class FileSubmissionService {
     private final FileSubmissionMapper mapper;
+    private final AssignmentMapper assignmentMapper;
     private final ObjectStorageService objectStorageService;
+
+    private void throwIfNotAllowedMultipleSubmission(Long assignmentId){
+        Assignment assignment = assignmentMapper.selectAssById(assignmentId);
+        if(!assignment.getAllowUpdateSubmission()){ // if not allowed, throw exception
+            throw new ExternalException(Response.getBadRequest("multiple-submission-not-allowed"));
+        }
+    }
 
     @Transactional
     public void submitFile(FileSubmission fileSubmission, MultipartFile file) throws IOException {
+        // check if have previous submission, if does, check if multiple submission valid.
+        Long assignmentId = fileSubmission.getAssignmentId(),studentId = fileSubmission.getStudentId();
+        FileSubmission prevSubmission= mapper.selectPreviousSubmission(assignmentId,studentId);
+        if(prevSubmission!=null) { // has previous
+            ///check if allowed multiple submission
+            throwIfNotAllowedMultipleSubmission(assignmentId);
+            // if allowed, remove old
+            fileSubmission.setFileSubmissionId(prevSubmission.getFileSubmissionId());
+            objectStorageService.delete(prevSubmission.getFileName());
+        }
         fileSubmission.setFileName(
                 UUID.randomUUID().toString() + file.getSize());
         mapper.insert(fileSubmission);
-        objectStorageService.save(fileSubmission.getFileName(),file.getBytes());
+        objectStorageService.save(fileSubmission.getFileName(), file.getBytes());
     }
     @Transactional
     public void deleteFileSubmission(Long id) {
         var fileSubmission = mapper.selectById(id);
+        if(fileSubmission==null){ // 幂等性
+            return;
+        }
+        throwIfNotAllowedMultipleSubmission(fileSubmission.getAssignmentId());
         var name = fileSubmission.getFileName();
         mapper.deleteById(id);
         objectStorageService.delete(name);
